requirements.txt
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.1
requests==2.32.3
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.10.0
tzlocal==5.2
Unidecode==1.3.8
urllib3==2.2.3
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
